requirements.txt
mysql-connector-python
bcrypt
pytest
configparser